basically the rest of the program
Made another file to return color values to be added to the list of detectable values.

Made function that gets contours from detected colours and returns the tip of the shape to be drawn on.

Made function that takes all the points and colours added and draws them out onto a resulting image